pushnotify/requirements.txt
auth==0.5.3
backcall==0.1.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
decorator==4.2.1
Django==1.11
django-cors-headers==2.1.0
django-sslserver==0.19
django-uuidfield==0.5.0
djangorestframework==3.6.3
eventlet==0.22.1
falcon==1.4.1
fcm-django==0.2.10
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyfcm==1.3.1
Pygments==2.2.0
pymongo==3.6.1
python-mimeparse==1.6.0
pytz==2018.3
requests==2.18.4
requests-toolbelt==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
